Crate nioruntime_http

Structs

ConnData
Connection Data used internally It is held in a lock that is used to determine if the thread has panicked.

HttpConfig
The configuration struct for an HttpServer. The Default trait is implemented for this struct so default can be used like below.

HttpServer
The main struct representing an HttpServer.

WriteHandle
A handle that is associated with a particular connection and may be used for writing to the socket.

Enums

HttpMethod

HttpVersion